Add files via upload
This is my fence cost calculator and uses many of the skills we learnt from the turtorials. This program includes a number checker that checks the meausrments of the fence are more than 0, gets the width: length and cost per metre of the fence, calculates the perimetre, calculates the total cost that they will have to spend for their total perimetre and can be done more than once or can be ended. It also includes a nice message that wishes them a good day as well as an emoji to make it more friendly.